feedback-ui/requirements.txt
firebase-functions==0.1.*
requests==2.*
google-auth==2.*
firebase-admin==6.2.*
mesop
pandas
gunicorn